% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{competitions}
\alias{competitions}
\title{All competition urls}
\format{
A data frame with 17 columns and XX rows:
\describe{
\item{type_id}{type_id of the matchday}
\item{type}{type of the matchday}
\item{season_id}{season_id of the matchday}
\item{season}{season of the matchday}
\item{league_id}{league_id of the matchday}
\item{league}{league of the matchday}
\item{matchday_id}{matchday_id of the matchday}
\item{matchday}{matchday label of the matchday}
\item{matchday_url}{matchday url of the matchday}
\item{datetime}{Date and time of the competition as POSIXct}
\item{location}{Address of competition stadium}
\item{title}{Title of competition; generally the two teams}
\item{home_team}{Name of home team}
\item{guest_team}{Name of guest team}
\item{competition_url}{URL to competition page}
\item{score}{Official final score result of competition}
\item{gp}{Official final gp result of competition}
}
}
\source{
https://www.deutsche-turnliga.de/archiv.html?Typ=Mann&SaisonID=20&LigaID=760&TagesID=666
}
\usage{
competitions
}
\description{
All crawled matchdays binded together
}
\keyword{datasets}
