% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{matchday_urls}
\alias{matchday_urls}
\title{Dataset with all matchday urls}
\format{
A data frame with 5 columns and 519 rows:
\describe{
\item{type}{type of the matchday}
\item{season}{season of the matchday}
\item{league}{league of the matchday}
\item{matchday}{matchday label of the matchday}
\item{url}{url to matchday}
}
}
\usage{
matchday_urls
}
\description{
This data set contains urls to all match days in
currently parseable leagues.
}
\keyword{datasets}
