% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pairings}
\alias{pairings}
\title{All crawled pairings binded together}
\format{
A data frame with 16 columns and XX rows:
\describe{
\item{home_team}{Name of the home team}
\item{guest_team}{Name of the guest team}
\item{event}{Event of the pairing}
\item{pairing_order}{Order of pairings per event}
\item{home_gymnast}{Name of gymnast of home team}
\item{home_gymnast_url}{URL to profile of gymnast of home team}
\item{home_starts}{Logical value if the home team started the pairing}
\item{home_d_value}{D-value of the home gymnast}
\item{home_end_value}{End value of the home gymnast}
\item{home_score_value}{Scraped score result of the home gymnast}
\item{guest_gymnast}{Name of gymnast of guest team}
\item{guest_gymnast_url}{URL to profile of gymnast of guest team}
\item{guest_d_value}{D-value of the guest gymnast}
\item{guest_end_value}{End value of the guest gymnast}
\item{guest_score_value}{Scraped score result of the guest gymnast}
\item{competition_url}{Url for competition}
}
}
\source{
https://www.deutsche-turnliga.de/dtl/historie/archiv/detailsm0.html?ID=2764
}
\usage{
pairings
}
\description{
All crawled pairings binded together
}
\keyword{datasets}
