% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{routines}
\alias{routines}
\title{All crawled routines binded together}
\format{
A data frame with 16 columns and XX rows:
\describe{
\item{team}{Name of the team}
\item{event}{Event of the pairing}
\item{pairing_order}{Order of pairings per event}
\item{gymnast}{Name of gymnast of home team}
\item{gymnast_url}{URL to profile of gymnast of home team}
\item{starts}{Logical value if the routine started the pairing}
\item{is_home}{Logical value if the team is the home team}
\item{d_value}{D-value of the home gymnast}
\item{end_value}{End value of the home gymnast}
\item{score_value}{Scraped score result of the home gymnast}
\item{competition_url}{Url for competition}
}
}
\source{
https://www.deutsche-turnliga.de/dtl/historie/archiv/detailsm0.html?ID=2764
}
\usage{
routines
}
\description{
All crawled routines binded together
}
\keyword{datasets}
