% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition.R
\name{parse_competition}
\alias{parse_competition}
\title{Parse competition results
Parse or scrape results from the competition site}
\usage{
parse_competition(url)
}
\arguments{
\item{url}{page url}
}
\value{
data.frame
}
\description{
Parse competition results
Parse or scrape results from the competition site
}
