% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/competition.R
\name{parse_competition}
\alias{parse_competition}
\title{Parse competition results}
\usage{
parse_competition(url)
}
\arguments{
\item{url}{page url}
}
\value{
data.frame
}
\description{
Parse or scrape results from the competition site
}
\examples{
\dontrun{
parse_competition("https://www.deutsche-turnliga.de/dtl/historie/archiv/detailsm0.html?ID=2787")
}
}
