% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchday.R
\name{parse_matchday}
\alias{parse_matchday}
\title{Parse matchday data
Parse or scrape the meta data of a matchday from a matchday overview page.}
\usage{
parse_matchday(html)
}
\arguments{
\item{url}{page url}
}
\value{
data.frame
}
\description{
Parse matchday data
Parse or scrape the meta data of a matchday from a matchday overview page.
}
